Add ball collisions

+++ Ball.h
@@ -2,11 +2,11 @@
 
 class Ball {
 public:
-	int x, y;
+	float x, y;
 	int speedX, speedY;
 	float radius;
 
-	Ball(int x, int y, int speedX, int speedY, float radius);
+	Ball(float x, float y, int speedX, int speedY, float radius);
 	void Draw();
 	void Move();
 };
